--- src/Kaiko-Api-Client/Kaiko-Api-Client.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30128.74
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kaiko-Api-Client", "Kaiko-Api-Client\Kaiko-Api-Client.csproj", "{1249E157-A5E6-4658-8025-86CE230CA416}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1249E157-A5E6-4658-8025-86CE230CA416}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1249E157-A5E6-4658-8025-86CE230CA416}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1249E157-A5E6-4658-8025-86CE230CA416}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1249E157-A5E6-4658-8025-86CE230CA416}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1B8CE014-A9EF-4AC3-B0EE-BE721274FA40}
-	EndGlobalSection
-EndGlobal
