+++ src/Packable.Projects.props
@@ -1,0 +1,18 @@
+<Project>
+  <PropertyGroup>
+    <PackageIconUrl>./trakx-logo.svg</PackageIconUrl>
+    <IsPackable>true</IsPackable>
+    <PublishRepositoryUrl>true</PublishRepositoryUrl>
+    <EmbedUntrackedSources>true</EmbedUntrackedSources>
+    <IncludeSymbols>true</IncludeSymbols>
+    <PackageTags>trakx</PackageTags>
+    <PackageIcon>package-icon.png</PackageIcon>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+    <SignAssembly>true</SignAssembly>
+  </PropertyGroup>
+  <ItemGroup>
+    <None Include="$(MSBuildThisFileDirectory)package-icon.png" PackagePath="." Pack="true" />
+  </ItemGroup>
+</Project>
