+++ src/Trakx.Kaiko.ApiClient.sln
@@ -3,9 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30128.74
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trakx.Kaiko.ApiClient", "Kaiko-Api-Client\Trakx.Kaiko.ApiClient.csproj", "{1249E157-A5E6-4658-8025-86CE230CA416}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trakx.Kaiko.ApiClient", "Trakx.Kaiko.ApiClient\Trakx.Kaiko.ApiClient.csproj", "{1249E157-A5E6-4658-8025-86CE230CA416}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trakx.Kaiko.ApiClient.Tests", "Kaiko-Api-Client.Tests\Trakx.Kaiko.ApiClient.Tests.csproj", "{11085538-FCFD-461D-8B92-0A7E4FC6E69F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trakx.Kaiko.ApiClient.Tests", "Trakx.Kaiko.ApiClient.Tests\Trakx.Kaiko.ApiClient.Tests.csproj", "{11085538-FCFD-461D-8B92-0A7E4FC6E69F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionItems", "{F9183CBB-5B9E-4E27-A07E-261D68E1A66A}"
+	ProjectSection(SolutionItems) = preProject
+		.env = .env
+		Common.Projects.props = Common.Projects.props
+		Packable.Projects.props = Packable.Projects.props
+		package-icon.png = package-icon.png
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
