+++ src/Trakx.Kaiko.ApiClient/ClientConfigurator.cs
@@ -6,11 +6,8 @@ namespace Trakx.Kaiko.ApiClient
 {
     internal class ClientConfigurator
     {
-        private readonly IServiceProvider _provider;
-
         public ClientConfigurator(IServiceProvider provider)
         {
-            _provider = provider;
             Configuration = provider.GetService<IOptions<KaikoConfiguration>>().Value;
         }
