+++ src/Trakx.Kaiko.ApiClient/ClientRegistrations.tt
@@ -1,0 +1,51 @@
+﻿<#@ template debug="false" hostspecific="false" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.Linq" #>
+<#@ import namespace="System.Text" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ output extension=".cs" #>
+using System;
+using System.Net.Http;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Polly;
+using Polly.Contrib.WaitAndRetry;
+using Polly.Extensions.Http;
+
+
+namespace Trakx.Kaiko.ApiClient
+{
+    public static partial class AddKaikoClientExtension
+    {
+        private static void AddClients(this IServiceCollection services)
+        {
+            var delay = Backoff.DecorrelatedJitterBackoffV2(medianFirstRetryDelay: TimeSpan.FromMilliseconds(100), retryCount: 10, fastFirst: true);
+            <# 
+            var clientNames = new [] { "AggregatesClient", "ExchangesClient", "InstrumentsClient" };
+
+            foreach(var name in clientNames)
+            {
+            #>
+                        
+            services.AddHttpClient<I<#= name #>, <#= name #>>()
+                .AddPolicyHandler((s, request) => 
+                    Policy<HttpResponseMessage>.Handle<ApiException>()
+                    .Or<HttpRequestException>()
+                    .OrTransientHttpStatusCode()
+                    .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.NotFound)
+                    .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.Unauthorized)
+                    .OrResult(msg => msg.StatusCode == System.Net.HttpStatusCode.Forbidden)
+                    .WaitAndRetryAsync(delay,
+                        onRetry: (result, timeSpan, retryCount, context) =>
+                        {
+                            var logger = s.GetService<ILogger<<#= name #>>>();
+                            LogFailure(logger, result, timeSpan, retryCount, context);
+                        })
+                    .WithPolicyKey("<#= name #>"));
+
+        <#
+        }
+        #>
+}
+    }
+}
