+++ src/Trakx.Kaiko.ApiClient/MarketDataClients.cs
@@ -18,7 +18,7 @@ namespace Trakx.Kaiko.ApiClient
     using System = global::System;
     
     [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.8.2.0 (NJsonSchema v10.2.1.0 (Newtonsoft.Json v12.0.0.0))")]
-    public partial interface IMarketDataClient
+    public partial interface IAggregatesClient
     {
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <param name="commodity">The data commodity</param>
@@ -35,13 +35,13 @@ namespace Trakx.Kaiko.ApiClient
     }
     
     [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.8.2.0 (NJsonSchema v10.2.1.0 (Newtonsoft.Json v12.0.0.0))")]
-    internal partial class MarketDataClient : AuthorisedClient, IMarketDataClient
+    internal partial class AggregatesClient : AuthorisedClient, IAggregatesClient
     {
         private string _baseUrl = "https://eu.market-api.kaiko.io/";
         private System.Net.Http.HttpClient _httpClient;
         private System.Lazy<Newtonsoft.Json.JsonSerializerSettings> _settings;
     
-        public MarketDataClient(ClientConfigurator configuration, System.Net.Http.HttpClient httpClient) : base(configuration)
+        public AggregatesClient(ClientConfigurator configuration, System.Net.Http.HttpClient httpClient) : base(configuration)
         {
             _httpClient = httpClient; 
             _settings = new System.Lazy<Newtonsoft.Json.JsonSerializerSettings>(CreateSerializerSettings);
